fix(diet): 당류(sugar_g) 계약을 소수로 통일
프론트는 항목 당류를 double 로 다루고 음식 단위(food_nutrients.sugar_g)도
이미 Float 인데 항목 단위만 Integer 라, 실서버 경로에서 `sugar_g=8.5` 가
Pydantic int 검증에 걸려 422 로 거부됐다(6.3+8.5 합도 절삭).

- `diet_entries.sugar_g` Integer → Float (마이그레이션 0018)
- `diet.py`·`diet_api.py` 의 `sugar_g`/`total_sugar_g` → float
- `trainer_api.py` 회원 총 당류 → float (이슈의 "필요 시 검토" 항목)

컬럼이 소수가 되면서 당류가 흘러드는 응답 스키마도 함께 맞췄다. 안 고치면
Pydantic 이 소수를 int 로 못 넣어 대시보드가 500 난다:

- `DashboardIndicator.current` → float (당류 지표가 17.8 을 싣는다)
- `DashboardNutritionDay.sugar_g` → float, 집계 누산기 타입도 float
- 목표치(max)는 세 지표 모두 정수라 int 유지

float 로 넓히면서 NaN/inf 가 새지 않도록 `allow_inf_nan=False` 를 붙였다.

로컬 Postgres 임시 DB 에서 0017 → 0018 적용, 기존 정수 값 보존(8 → 8.0),
소수 저장·합계(6.3+8.5=14.8) 무절삭, downgrade·재적용까지 확인했다.
downgrade 는 소수부가 버려지므로 마이그레이션에 손실을 명시했다.

+++ backend/app/schemas/dashboard_api.py
@@ -9,7 +9,9 @@ from app.schemas.diet_api import Macros
 
 class DashboardIndicator(BaseModel):
     label: str            # 칼로리 | 나트륨 | 당류
-    current: int
+    # 당류가 소수(17.8g)라 current 는 float. 칼로리·나트륨은 정수 값이 그대로
+    # 실려 나가고, 목표치(max)는 세 지표 모두 정수라 int 를 유지한다.
+    current: float
     max: int
     unit: str
     over_budget: bool = False
@@ -29,7 +31,7 @@ class DashboardNutritionDay(BaseModel):
     label: str       # 요일 라벨(월/화/…) — 프론트 x축용
     calories: int
     sodium_mg: int
-    sugar_g: int
+    sugar_g: float
 
 
 class DashboardSummary(BaseModel):
